Enable event consumer.
