+++ frontend/vite.config.ts
@@ -1,5 +1,3 @@
-/// <reference types="vite/client" />
-
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react-swc';
 import { createHtmlPlugin } from 'vite-plugin-html';
